WebHost: prevent import of worlds subsystem when launching Process()
